Changed: rake juggernaut:install checks for jquery.js and installs the right javascript Added: first, non working jQuery plugin for juggernaut. We'll get there in time.

+++ install.rb
@@ -8,7 +8,15 @@ FileUtils.mkdir_p("#{there}/public/juggernaut/")
 
 puts "Installing Juggernaut..."
 FileUtils.cp("#{here}/media/swfobject.js", "#{there}/public/javascripts/juggernaut/")
-FileUtils.cp("#{here}/media/juggernaut.js", "#{there}/public/javascripts/juggernaut/")
+puts "Checking for \"jquery.js\" in \"#{RAILS_ROOT}/public/javascripts/\"..."
+if File.exist?("#{RAILS_ROOT}/public/javascripts/jquery.js")
+  puts "Using jQuery? Right on!"
+  FileUtils.cp("#{here}/media/jquery.juggernaut.js", "#{there}/public/javascripts/juggernaut/")
+else
+  puts "Still on Prototype? Oh well..."
+  FileUtils.cp("#{here}/media/juggernaut.js", "#{there}/public/javascripts/juggernaut/")
+end 
+  
 FileUtils.cp("#{here}/media/juggernaut.swf", "#{there}/public/juggernaut/")
 FileUtils.cp("#{here}/media/expressinstall.swf", "#{there}/public/juggernaut/")
